update kerusakan & perbaiakan

+++ application/controllers/Pengembalian.php
@@ -103,14 +103,6 @@ class Pengembalian extends MY_Controller
 
             $nama = encrypt_url($this->input->post('nama'));
             $config['upload_path']   = './assets/foto/kembali/';
-            if($kondisi=='2'){
-                $config['upload_path']   = './assets/foto/kerusakan_alat/';
-            }
-
-            if($kondisi=='1'){
-                $config['upload_path']   = './assets/foto/perbaikan_alat/';
-            }
-            
             $config['allowed_types'] = 'gif|jpg|jpeg|png'; //mencegah upload backdor
             $config['max_size']      = '1000';
             $config['max_width']     = '2000';
@@ -120,8 +112,8 @@ class Pengembalian extends MY_Controller
             $this->upload->initialize($config);
 
             if ($this->upload->do_upload('imagefile')){
-             $gambar = $this->upload->data();
-             $save  = array(
+               $gambar = $this->upload->data();
+               $save  = array(
                 'nama'         => htmlspecialchars_decode(ucwords($this->input->post('nama'))),
                 'id_jabatan'    => $this->input->post('id_jabatan'),
                 'id_peminjaman'    => $this->input->post('id_peminjaman'),
@@ -137,11 +129,11 @@ class Pengembalian extends MY_Controller
                 'foto'         => $gambar['file_name'],
 
             );
-             $this->Mod_pengembalian->insert("pengembalian", $save);
-             $id_pengembalian = $this->db->insert_id();
-             echo json_encode(array("status" => TRUE));
+               $this->Mod_pengembalian->insert("pengembalian", $save);
+               $id_pengembalian = $this->db->insert_id();
+               echo json_encode(array("status" => TRUE));
 
-             if ($kondisi=='2') {
+               if ($kondisi=='2') {
                 $save1  = array(
                     'id_alat'    => $this->input->post('id_alat'),
                     'id_satuan'    => $this->input->post('id_satuan'),
@@ -155,6 +147,9 @@ class Pengembalian extends MY_Controller
                     'foto'         => $gambar['file_name'],
 
                 );
+                $path= './assets/foto/kembali/'.$gambar['file_name'];
+                $path1= './assets/foto/kerusakan_alat/'.$gambar['file_name'];
+                copy($path, $path1);
                 $this->Mod_pengembalian->insert("kerusakan_alat", $save1);
             }
 
@@ -172,6 +167,9 @@ class Pengembalian extends MY_Controller
                     'foto'         => $gambar['file_name'],
 
                 );
+                $path= './assets/foto/kembali/'.$gambar['file_name'];
+                $path1= './assets/foto/perbaikan_alat/'.$gambar['file_name'];
+                copy($path, $path1);
                 $this->Mod_pengembalian->insert("perbaikan_alat", $save);
 
             }
@@ -219,7 +217,8 @@ class Pengembalian extends MY_Controller
 }
 
 public function update()
-{
+{   
+    $id      = $this->input->post('id');
     $kondisi = $this->input->post('id_kondisi');
     if($kondisi=='2'){
         $this->_validate1();
@@ -234,6 +233,9 @@ public function update()
     $stok_in1 = (!empty($cek->stok_in)) ? $cek->stok_in : '0' ;
     $stin = $stok_in+$stok_in1;
 
+    $r=$this->Mod_pengembalian->get($id);
+    $kondisi_s=$r->id_kondisi;
+
     if ($stok_out==$stin) {
         $save1 = array('status' => '1', );
         $this->Mod_peminjaman->update($id_peminjaman, $save1);
@@ -243,7 +245,7 @@ public function update()
     }
 
 
-    $id      = $this->input->post('id');
+    
     if(!empty($_FILES['imagefile']['name'])) {
         // $this->_validate();
         $id = $this->input->post('id_user');
@@ -251,13 +253,7 @@ public function update()
         $nama = encrypt_url($this->input->post('nama'));
         $config['upload_path']   = './assets/foto/kembali/';
             $config['allowed_types'] = 'gif|jpg|jpeg|png'; //mencegah upload backdor
-            if($kondisi=='2'){
-                $config['upload_path']   = './assets/foto/kerusakan_alat/';
-            }
 
-            if($kondisi=='1'){
-                $config['upload_path']   = './assets/foto/perbaikan_alat/';
-            }
             $config['max_size']      = '1000';
             $config['max_width']     = '2000';
             $config['max_height']    = '1024';
@@ -266,8 +262,8 @@ public function update()
             $this->upload->initialize($config);
 
             if ($this->upload->do_upload('imagefile')){
-             $gambar = $this->upload->data();
-             $save  = array(
+               $gambar = $this->upload->data();
+               $save  = array(
                 'nama'         => htmlspecialchars_decode(ucwords($this->input->post('nama'))),
                 'id_jabatan'    => $this->input->post('id_jabatan'),
                 'id_alat'    => $this->input->post('id_alat'),
@@ -280,20 +276,21 @@ public function update()
                 'foto'         => $gambar['file_name'],
 
             );
-             $g = $this->Mod_pengembalian->getImage($id)->row();
+               $g = $this->Mod_pengembalian->getImage($id)->row();
 
-             if (!empty($g->foto)) {
+               if (!empty($g->foto) && $g->foto!=NULL) {
                 //hapus gambar yg ada diserver
-                unlink('assets/foto/kembali/'.$g->foto);
-                if($kondisi=='2'){
+               
+                if($kondisi_s=='1'){
 
                     unlink('./assets/foto/perbaikan_alat/'.$g->foto);
-                }elseif($kondisi=='2'){
+                }elseif($kondisi_s=='2'){
 
                     unlink('./assets/foto/kerusakan_alat/'.$g->foto);
                 }else{
                     unlink('./assets/foto/perbaikan_alat/'.$g->foto);
                     unlink('./assets/foto/kerusakan_alat/'.$g->foto);
+                     unlink('./assets/foto/kembali/'.$g->foto);
                 }
             }
             $this->Mod_pengembalian->update($id, $save);
@@ -304,14 +301,23 @@ public function update()
                     'id_alat'    => $this->input->post('id_alat'),
                     'id_satuan'    => $this->input->post('id_satuan'),
                     'id_kondisi'    => $this->input->post('id_kondisi'),
-                    'tgl_masuk'    => $this->input->post('tgl_in'),
-                    'stok'    => $this->input->post('stok_in'),
+                    'tgl_input'    => $this->input->post('tgl_in'),
+                    'stok_out'    => $this->input->post('stok_in'),
                     'keterangan'    => $this->input->post('keterangan'),
                     'id_user'    => $id_user,
                     'id_jurusan' => $id_jurusan,
+                     'id_pengembalian' => $id
 
                 );
-                $this->Mod_pengembalian->update_rusak($id, $save);//update kerusakan
+                $path= './assets/foto/kembali/'.$gambar['file_name'];
+                $path1= './assets/foto/kerusakan_alat/'.$gambar['file_name'];
+                copy($path, $path1);
+                 $cek_kerusakan=$this->Mod_pengembalian->cek_kerusakan_alat($id)->result();
+                if (count($cek_kerusakan)=='0') {
+                   $this->Mod_pengembalian->insert("kerusakan_alat", $save);
+                } else {
+                    $this->Mod_pengembalian->update_kerusakan($id, $save);
+                }
                 $this->Mod_pengembalian->delete_perbaikan($id, 'perbaikan_alat');  //hapus perbaikan
             }elseif ($kondisi=='1') {//jika maintenance
                 $save  = array(
@@ -323,9 +329,18 @@ public function update()
                     'keterangan'    => $this->input->post('keterangan'),
                     'id_user'    => $id_user,
                     'id_jurusan' => $id_jurusan,
+                    'id_pengembalian' => $id
 
                 );
-                $this->Mod_pengembalian->update_perbaikan($id, $save);//update perbaikan
+                $path= './assets/foto/kembali/'.$gambar['file_name'];
+                $path1= './assets/foto/perbaikan_alat/'.$gambar['file_name'];
+                copy($path, $path1);
+                 $cek_perbaikan=$this->Mod_pengembalian->cek_perbaikan_alat($id)->result();
+                 if (count($cek_perbaikan)=='0') {
+                     $this->Mod_pengembalian->insert("perbaikan_alat", $save);
+                 } else {
+                     $this->Mod_pengembalian->update_perbaikan($id, $save);
+                 }
                 $this->Mod_pengembalian->delete_kerusakan($id, 'kerusakan_alat'); //hapus kerusakan
             }else{
                  $this->Mod_pengembalian->delete_perbaikan($id, 'perbaikan_alat');  //hapus perbaikan
@@ -345,25 +360,79 @@ public function update()
             'keterangan'    => $this->input->post('keterangan'),
 
         );
-
+        
         $this->Mod_pengembalian->update($id, $save);
         echo json_encode(array("status" => TRUE));
-        if ($kondisi=='1') {
-            $save  = array(
-                'id_alat'    => $this->input->post('id_alat'),
-                'id_satuan'    => $this->input->post('id_satuan'),
-                'id_kondisi'    => $this->input->post('id_kondisi'),
-                'tgl_masuk'    => $this->input->post('tgl_in'),
-                'stok'    => $this->input->post('stok_in'),
-                'keterangan'    => $this->input->post('keterangan'),
-                'id_user'    => $id_user,
-                'id_jurusan' => $id_jurusan,
-                'id_pengembalian' => $id,
+         if ($kondisi=='2') {//Jika kondisi rusak 
+                $save  = array(
+                    'id_alat'    => $this->input->post('id_alat'),
+                    'id_satuan'    => $this->input->post('id_satuan'),
+                    'id_kondisi'    => $this->input->post('id_kondisi'),
+                    'tgl_input'    => $this->input->post('tgl_in'),
+                    'stok_out'    => $this->input->post('stok_in'),
+                    'keterangan'    => $this->input->post('keterangan'),
+                    'id_user'    => $id_user,
+                    'id_jurusan' => $id_jurusan,
+                     'id_pengembalian' => $id
 
-            );
-            $this->Mod_pengembalian->insert("perbaikan_alat", $save);
-            $this->Mod_pengembalian->delete_kerusakan($id, 'kerusakan_alat'); //hapus kerusakan
-        }
+                );
+                $g = $this->Mod_pengembalian->getImage($id)->row();
+
+                if (!empty($g->foto) && $g->foto!=NULL) {
+                //hapus gambar yg ada diserver
+                    unlink('./assets/foto/kembali/'.$g->foto);
+                    if ($g->id_kondisi=='1') {
+                        unlink('assets/foto/perbaikan_alat/'.$g->foto);
+                    } 
+                    if ($g->id_kondisi=='2') {
+                        unlink('./assets/foto/kerusakan_alat/'.$g->foto);
+                    }
+
+
+
+                }
+                
+                $path= './assets/foto/kembali/'.$g->foto;
+                $path1= './assets/foto/kerusakan_alat/'.$g->foto;
+                copy($path, $path1);
+                $cek_kerusakan=$this->Mod_pengembalian->cek_kerusakan_alat($id)->result();
+                if (count($cek_kerusakan)=='0') {
+                   $this->Mod_pengembalian->insert("kerusakan_alat", $save);
+                } else {
+                    $this->Mod_pengembalian->update_kerusakan($id, $save);
+                }
+                
+                 
+                $this->Mod_pengembalian->delete_perbaikan($id, 'perbaikan_alat');  //hapus perbaikan
+            }elseif ($kondisi=='1') {//jika maintenance
+                $save  = array(
+                    'id_alat'    => $this->input->post('id_alat'),
+                    'id_satuan'    => $this->input->post('id_satuan'),
+                    'id_kondisi'    => $this->input->post('id_kondisi'),
+                    'tgl_masuk'    => $this->input->post('tgl_in'),
+                    'stok'    => $this->input->post('stok_in'),
+                    'keterangan'    => $this->input->post('keterangan'),
+                    'id_user'    => $id_user,
+                    'id_jurusan' => $id_jurusan,
+                    'id_pengembalian' => $id
+
+                );
+                $path= './assets/foto/kembali/'.$g->foto;
+                $path1= './assets/foto/perbaikan_alat/'.$g->foto;
+                copy($path, $path1);
+                 $cek_perbaikan=$this->Mod_pengembalian->cek_perbaikan_alat($id)->result();
+                 if (count($cek_perbaikan)=='0') {
+                     $this->Mod_pengembalian->insert("perbaikan_alat", $save);
+                 } else {
+                     $this->Mod_pengembalian->update_perbaikan($id, $save);
+                 }
+                 
+                 
+                $this->Mod_pengembalian->delete_kerusakan($id, 'kerusakan_alat'); //hapus kerusakan
+            }else{
+                 $this->Mod_pengembalian->delete_perbaikan($id, 'perbaikan_alat');  //hapus perbaikan
+                $this->Mod_pengembalian->delete_kerusakan($id, 'kerusakan_alat'); //hapus kerusakan
+            }
 
 
     }
@@ -390,16 +459,16 @@ public function delete()
     $this->Mod_peminjaman->update($id_peminjaman, $save1);
     $this->Mod_pengembalian->delete_perbaikan($id, 'perbaikan_alat');  //hapus perbaikan
      $this->Mod_pengembalian->delete_kerusakan($id, 'kerusakan_alat'); //hapus kerusakan
-    $g = $this->Mod_pengembalian->getImage($id)->row();
+     $g = $this->Mod_pengembalian->getImage($id)->row();
 
-    if (!empty($g->foto) || $g->foto != NULL) {
+     if (!empty($g->foto) && $g->foto!=NULL) {
                 //hapus gambar yg ada diserver
-        unlink('assets/foto/kembali/'.$g->foto);
+        unlink('./assets/foto/kembali/'.$g->foto);
         if ($g->id_kondisi=='1') {
             unlink('assets/foto/perbaikan_alat/'.$g->foto);
         } 
-        if ($g->id_kondisi=='1') {
-            unlink('assets/foto/kerusakan_alat/'.$g->foto);
+        if ($g->id_kondisi=='2') {
+            unlink('./assets/foto/kerusakan_alat/'.$g->foto);
         }
